MyDoc App Backend - FastAPI/Servers(backend)/requirements.txt
fastapi
uvicorn
mysql-connector-python
mysqlclient
sqlalchemy
passlib
bcrypt
python-jose
httpx
python-multipart